+++ src/db/oci/Schema.php
@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 /**
  * @link https://www.yiiframework.com/
  * @copyright Copyright (c) 2008 Yii Software LLC
@@ -8,6 +10,8 @@
 
 namespace yii\db\oci;
 
+use Exception;
+use PDO;
 use Yii;
 use yii\base\InvalidCallException;
 use yii\base\NotSupportedException;
@@ -23,8 +27,19 @@ use yii\db\oci\ColumnSchema;
 use yii\db\TableSchema;
 use yii\helpers\ArrayHelper;
 
+use function array_change_key_case;
+use function array_keys;
+use function array_merge;
+use function count;
+use function explode;
+use function implode;
+use function str_replace;
+use function strpos;
+use function strtoupper;
+use function trim;
+
 /**
- * Schema is the class for retrieving metadata from an Oracle database.
+ * Schema is the class for retrieving metadata from an Oracle database (version 12c and later).
  *
  * @property-read string $lastInsertID The row ID of the last row inserted, or the last value retrieved from
  * the sequence object.
@@ -68,7 +83,7 @@ class Schema extends \yii\db\Schema implements ConstraintFinderInterface
             $username = $this->db->username;
 
             if (empty($username)) {
-                $username = isset($this->db->masters[0]['username']) ? $this->db->masters[0]['username'] : '';
+                $username = $this->db->masters[0]['username'] ?? '';
             }
 
             $this->defaultSchema = strtoupper($username);
@@ -77,83 +92,70 @@ class Schema extends \yii\db\Schema implements ConstraintFinderInterface
 
     /**
      * {@inheritdoc}
-     */
-    protected function resolveTableName($name)
-    {
-        $parts = explode('.', str_replace('"', '', $name));
-
-        $tableName = $parts[1] ?? $parts[0];
-        $schemaName = isset($parts[1]) ? $parts[0] : $this->defaultSchema;
-
-        $fullName = $schemaName !== $this->defaultSchema
-            ? "{$schemaName}.{$tableName}"
-            : $tableName;
-
-        $resolvedName = new TableSchema();
-
-        $resolvedName->name = $tableName;
-        $resolvedName->schemaName = $schemaName;
-        $resolvedName->fullName = $fullName;
-
-        return $resolvedName;
-    }
-
-    /**
-     * {@inheritdoc}
-     * @see https://docs.oracle.com/en/database/oracle/oracle-database/19/dbseg/managing-security-for-oracle-database-users.html
+     *
+     * @see https://docs.oracle.com/en/database/oracle/oracle-database/19/refrn/ALL_USERS.html
      */
     protected function findSchemaNames()
     {
-        static $sql = <<<'SQL'
-SELECT "u"."USERNAME"
-FROM "DBA_USERS" "u"
-WHERE "u"."DEFAULT_TABLESPACE" NOT IN ('SYSTEM', 'SYSAUX')
-ORDER BY "u"."USERNAME" ASC
-SQL;
+        $sql = <<<SQL
+            SELECT "u"."USERNAME"
+            FROM "ALL_USERS" "u"
+            WHERE "u"."ORACLE_MAINTAINED" = 'N'
+            ORDER BY "u"."USERNAME" ASC
+            SQL;
 
         return $this->db->createCommand($sql)->queryColumn();
     }
 
     /**
      * {@inheritdoc}
+     *
+     * @see https://docs.oracle.com/en/database/oracle/oracle-database/21/refrn/ALL_TABLES.html
+     * @see https://docs.oracle.com/en/database/oracle/oracle-database/19/refrn/ALL_VIEWS.html
+     * @see https://docs.oracle.com/en/database/oracle/oracle-database/19/refrn/ALL_MVIEWS.html
      */
     protected function findTableNames($schema = '')
     {
         if ($schema === '') {
-            $sql = <<<'SQL'
-SELECT
-    TABLE_NAME
-FROM USER_TABLES
-UNION ALL
-SELECT
-    VIEW_NAME AS TABLE_NAME
-FROM USER_VIEWS
-UNION ALL
-SELECT
-    MVIEW_NAME AS TABLE_NAME
-FROM USER_MVIEWS
-ORDER BY TABLE_NAME
-SQL;
+            $sql = <<<SQL
+                SELECT "t"."TABLE_NAME"
+                FROM "USER_TABLES" "t"
+                UNION ALL
+                SELECT "v"."VIEW_NAME" AS "TABLE_NAME"
+                FROM "USER_VIEWS" "v"
+                UNION ALL
+                SELECT "m"."MVIEW_NAME" AS "TABLE_NAME"
+                FROM "USER_MVIEWS" "m"
+                ORDER BY "TABLE_NAME"
+                SQL;
             $command = $this->db->createCommand($sql);
         } else {
-            $sql = <<<'SQL'
-SELECT
-    OBJECT_NAME AS TABLE_NAME
-FROM ALL_OBJECTS
-WHERE
-    OBJECT_TYPE IN ('TABLE', 'VIEW', 'MATERIALIZED VIEW')
-    AND OWNER = :schema
-ORDER BY OBJECT_NAME
-SQL;
+            $sql = <<<SQL
+                SELECT "t"."TABLE_NAME"
+                FROM "ALL_TABLES" "t"
+                WHERE "t"."OWNER" = :schema
+                UNION ALL
+                SELECT "v"."VIEW_NAME" AS "TABLE_NAME"
+                FROM "ALL_VIEWS" "v"
+                WHERE "v"."OWNER" = :schema
+                UNION ALL
+                SELECT "m"."MVIEW_NAME" AS "TABLE_NAME"
+                FROM "ALL_MVIEWS" "m"
+                WHERE "m"."OWNER" = :schema
+                ORDER BY "TABLE_NAME"
+                SQL;
             $command = $this->db->createCommand($sql, [':schema' => $schema]);
         }
 
         $rows = $command->queryAll();
+
         $names = [];
+
         foreach ($rows as $row) {
-            if ($this->db->slavePdo->getAttribute(\PDO::ATTR_CASE) === \PDO::CASE_LOWER) {
+            if ($this->db->slavePdo->getAttribute(PDO::ATTR_CASE) === PDO::CASE_LOWER) {
                 $row = array_change_key_case($row, CASE_UPPER);
             }
+
             $names[] = $row['TABLE_NAME'];
         }
 
@@ -200,41 +202,54 @@ SQL;
 
     /**
      * {@inheritdoc}
+     *
+     * @see https://docs.oracle.com/en/database/oracle/oracle-database/19/refrn/USER_INDEXES.html
+     * @see https://docs.oracle.com/en/database/oracle/oracle-database/19/refrn/USER_IND_COLUMNS.html
+     * @see https://docs.oracle.com/en/database/oracle/oracle-database/19/refrn/USER_CONSTRAINTS.html
      */
     protected function loadTableIndexes($tableName)
     {
-        static $sql = <<<'SQL'
-SELECT
-    /*+ PUSH_PRED("ui") PUSH_PRED("uicol") PUSH_PRED("uc") */
-    "ui"."INDEX_NAME" AS "name",
-    "uicol"."COLUMN_NAME" AS "column_name",
-    CASE "ui"."UNIQUENESS" WHEN 'UNIQUE' THEN 1 ELSE 0 END AS "index_is_unique",
-    CASE WHEN "uc"."CONSTRAINT_NAME" IS NOT NULL THEN 1 ELSE 0 END AS "index_is_primary"
-FROM "SYS"."USER_INDEXES" "ui"
-LEFT JOIN "SYS"."USER_IND_COLUMNS" "uicol"
-    ON "uicol"."INDEX_NAME" = "ui"."INDEX_NAME"
-LEFT JOIN "SYS"."USER_CONSTRAINTS" "uc"
-    ON "uc"."OWNER" = "ui"."TABLE_OWNER" AND "uc"."CONSTRAINT_NAME" = "ui"."INDEX_NAME" AND "uc"."CONSTRAINT_TYPE" = 'P'
-WHERE "ui"."TABLE_OWNER" = :schemaName AND "ui"."TABLE_NAME" = :tableName
-    AND "ui"."INDEX_TYPE" != 'LOB'
-ORDER BY "uicol"."COLUMN_POSITION" ASC
-SQL;
+        $sql = <<<SQL
+            SELECT
+                "ui"."INDEX_NAME" AS "name",
+                "uicol"."COLUMN_NAME" AS "column_name",
+                CASE "ui"."UNIQUENESS" WHEN 'UNIQUE' THEN 1 ELSE 0 END AS "index_is_unique",
+                CASE WHEN "uc"."CONSTRAINT_NAME" IS NOT NULL THEN 1 ELSE 0 END AS "index_is_primary"
+            FROM "USER_INDEXES" "ui"
+            LEFT JOIN "USER_IND_COLUMNS" "uicol"
+                ON "uicol"."INDEX_NAME" = "ui"."INDEX_NAME"
+            LEFT JOIN "USER_CONSTRAINTS" "uc"
+                ON "uc"."OWNER" = "ui"."TABLE_OWNER"
+                AND "uc"."CONSTRAINT_NAME" = "ui"."INDEX_NAME"
+                AND "uc"."CONSTRAINT_TYPE" = 'P'
+            WHERE "ui"."TABLE_OWNER" = :schemaName
+                AND "ui"."TABLE_NAME" = :tableName
+                AND "ui"."INDEX_TYPE" != 'LOB'
+            ORDER BY "uicol"."COLUMN_POSITION" ASC
+            SQL;
 
         $resolvedName = $this->resolveTableName($tableName);
-        $indexes = $this->db->createCommand($sql, [
-            ':schemaName' => $resolvedName->schemaName,
-            ':tableName' => $resolvedName->name,
-        ])->queryAll();
+        $indexes = $this->db->createCommand(
+            $sql,
+            [
+                ':schemaName' => $resolvedName->schemaName,
+                ':tableName' => $resolvedName->name,
+            ],
+        )->queryAll();
         $indexes = $this->normalizePdoRowKeyCase($indexes, true);
         $indexes = ArrayHelper::index($indexes, null, 'name');
+
         $result = [];
+
         foreach ($indexes as $name => $index) {
-            $result[] = new IndexConstraint([
-                'isPrimary' => (bool) $index[0]['index_is_primary'],
-                'isUnique' => (bool) $index[0]['index_is_unique'],
-                'name' => $name,
-                'columnNames' => ArrayHelper::getColumn($index, 'column_name'),
-            ]);
+            $result[] = new IndexConstraint(
+                [
+                    'isPrimary' => (bool) $index[0]['index_is_primary'],
+                    'isUnique' => (bool) $index[0]['index_is_unique'],
+                    'name' => $name,
+                    'columnNames' => ArrayHelper::getColumn($index, 'column_name'),
+                ],
+            );
         }
 
         return $result;
@@ -299,41 +314,53 @@ SQL;
 
     /**
      * Collects the table column metadata.
+     *
      * @param TableSchema $table the table schema
      * @return bool whether the table exists
+     *
+     * @see https://docs.oracle.com/en/database/oracle/oracle-database/19/refrn/ALL_TAB_COLUMNS.html
+     * @see https://docs.oracle.com/en/database/oracle/oracle-database/19/refrn/ALL_OBJECTS.html
+     * @see https://docs.oracle.com/en/database/oracle/oracle-database/19/refrn/ALL_COL_COMMENTS.html
      */
     protected function findColumns($table)
     {
-        $sql = <<<'SQL'
-SELECT
-    A.COLUMN_NAME,
-    A.DATA_TYPE,
-    A.DATA_PRECISION,
-    A.DATA_SCALE,
-    (
-      CASE A.CHAR_USED WHEN 'C' THEN A.CHAR_LENGTH
-        ELSE A.DATA_LENGTH
-      END
-    ) AS DATA_LENGTH,
-    A.NULLABLE,
-    A.DATA_DEFAULT,
-    COM.COMMENTS AS COLUMN_COMMENT
-FROM ALL_TAB_COLUMNS A
-    INNER JOIN ALL_OBJECTS B ON B.OWNER = A.OWNER AND LTRIM(B.OBJECT_NAME) = LTRIM(A.TABLE_NAME)
-    LEFT JOIN ALL_COL_COMMENTS COM ON (A.OWNER = COM.OWNER AND A.TABLE_NAME = COM.TABLE_NAME AND A.COLUMN_NAME = COM.COLUMN_NAME)
-WHERE
-    A.OWNER = :schemaName
-    AND B.OBJECT_TYPE IN ('TABLE', 'VIEW', 'MATERIALIZED VIEW')
-    AND B.OBJECT_NAME = :tableName
-ORDER BY A.COLUMN_ID
-SQL;
+        $sql = <<<SQL
+            SELECT
+                "a"."COLUMN_NAME",
+                "a"."DATA_TYPE",
+                "a"."DATA_PRECISION",
+                "a"."DATA_SCALE",
+                (
+                    CASE "a"."CHAR_USED" WHEN 'C' THEN "a"."CHAR_LENGTH"
+                        ELSE "a"."DATA_LENGTH"
+                    END
+                ) AS "DATA_LENGTH",
+                "a"."NULLABLE",
+                "a"."DATA_DEFAULT",
+                "com"."COMMENTS" AS "COLUMN_COMMENT"
+            FROM "ALL_TAB_COLUMNS" "a"
+            INNER JOIN "ALL_OBJECTS" "b"
+                ON "b"."OWNER" = "a"."OWNER"
+                AND "b"."OBJECT_NAME" = "a"."TABLE_NAME"
+            LEFT JOIN "ALL_COL_COMMENTS" "com"
+                ON "a"."OWNER" = "com"."OWNER"
+                AND "a"."TABLE_NAME" = "com"."TABLE_NAME"
+                AND "a"."COLUMN_NAME" = "com"."COLUMN_NAME"
+            WHERE "a"."OWNER" = :schemaName
+                AND "b"."OBJECT_TYPE" IN ('TABLE', 'VIEW', 'MATERIALIZED VIEW')
+                AND "b"."OBJECT_NAME" = :tableName
+            ORDER BY "a"."COLUMN_ID"
+            SQL;
 
         try {
-            $columns = $this->db->createCommand($sql, [
-                ':tableName' => $table->name,
-                ':schemaName' => $table->schemaName,
-            ])->queryAll();
-        } catch (\Exception $e) {
+            $columns = $this->db->createCommand(
+                $sql,
+                [
+                    ':tableName' => $table->name,
+                    ':schemaName' => $table->schemaName,
+                ],
+            )->queryAll();
+        } catch (Exception $e) {
             return false;
         }
 
@@ -342,10 +369,12 @@ SQL;
         }
 
         foreach ($columns as $column) {
-            if ($this->db->slavePdo->getAttribute(\PDO::ATTR_CASE) === \PDO::CASE_LOWER) {
+            if ($this->db->slavePdo->getAttribute(PDO::ATTR_CASE) === PDO::CASE_LOWER) {
                 $column = array_change_key_case($column, CASE_UPPER);
             }
+
             $c = $this->createColumn($column);
+
             $table->columns[$c->name] = $c;
         }
 
@@ -358,20 +387,27 @@ SQL;
      * @param string $tableName
      * @internal param \yii\db\TableSchema $table->name the table schema
      * @return string|null whether the sequence exists
+     *
+     * @see https://docs.oracle.com/en/database/oracle/oracle-database/19/refrn/USER_DEPENDENCIES.html
+     * @see https://docs.oracle.com/en/database/oracle/oracle-database/19/refrn/USER_TRIGGERS.html
      */
     protected function getTableSequenceName($tableName)
     {
-        $sequenceNameSql = <<<'SQL'
-SELECT
-    UD.REFERENCED_NAME AS SEQUENCE_NAME
-FROM USER_DEPENDENCIES UD
-    JOIN USER_TRIGGERS UT ON (UT.TRIGGER_NAME = UD.NAME)
-WHERE
-    UT.TABLE_NAME = :tableName
-    AND UD.TYPE = 'TRIGGER'
-    AND UD.REFERENCED_TYPE = 'SEQUENCE'
-SQL;
-        $sequenceName = $this->db->createCommand($sequenceNameSql, [':tableName' => $tableName])->queryScalar();
+        $sequenceNameSql = <<<SQL
+            SELECT "ud"."REFERENCED_NAME" AS "SEQUENCE_NAME"
+            FROM "USER_DEPENDENCIES" "ud"
+            INNER JOIN "USER_TRIGGERS" "ut"
+                ON "ut"."TRIGGER_NAME" = "ud"."NAME"
+            WHERE "ut"."TABLE_NAME" = :tableName
+                AND "ud"."TYPE" = 'TRIGGER'
+                AND "ud"."REFERENCED_TYPE" = 'SEQUENCE'
+            SQL;
+
+        $sequenceName = $this->db->createCommand(
+            $sequenceNameSql,
+            [':tableName' => $tableName],
+        )->queryScalar();
+
         return $sequenceName === false ? null : $sequenceName;
     }
 
@@ -389,9 +425,14 @@ SQL;
         if ($this->db->isActive) {
             // get the last insert id from the master connection
             $sequenceName = $this->quoteSimpleTableName($sequenceName);
-            return $this->db->useMaster(function (Connection $db) use ($sequenceName) {
-                return $db->createCommand("SELECT {$sequenceName}.CURRVAL FROM DUAL")->queryScalar();
-            });
+
+            $sql = <<<SQL
+            SELECT {$sequenceName}.CURRVAL FROM DUAL
+            SQL;
+
+            return $this->db->useMaster(
+                static fn(Connection $db) => $db->createCommand($sql)->queryScalar(),
+            );
         } else {
             throw new InvalidCallException('DB Connection is not active.');
         }
@@ -406,16 +447,30 @@ SQL;
     protected function createColumn($column)
     {
         $c = $this->createColumnSchema();
+
         $c->name = $column['COLUMN_NAME'];
         $c->allowNull = $column['NULLABLE'] === 'Y';
         $c->comment = $column['COLUMN_COMMENT'] === null ? '' : $column['COLUMN_COMMENT'];
         $c->isPrimaryKey = false;
-        $this->extractColumnType($c, $column['DATA_TYPE'], $column['DATA_PRECISION'], $column['DATA_SCALE'], $column['DATA_LENGTH']);
-        $this->extractColumnSize($c, $column['DATA_TYPE'], $column['DATA_PRECISION'], $column['DATA_SCALE'], $column['DATA_LENGTH']);
+
+        $this->extractColumnType(
+            $c,
+            $column['DATA_TYPE'],
+            $column['DATA_PRECISION'],
+            $column['DATA_SCALE'],
+            $column['DATA_LENGTH'],
+        );
+        $this->extractColumnSize(
+            $c,
+            $column['DATA_TYPE'],
+            $column['DATA_PRECISION'],
+            $column['DATA_SCALE'],
+            $column['DATA_LENGTH'],
+        );
 
         $c->phpType = $this->getColumnPhpType($c);
 
-        // Store raw default for deferred resolution in loadTableSchema(), where isPrimaryKey is known.
+        // store raw default for deferred resolution in `loadTableSchema()`, where `isPrimaryKey` is known
         $c->defaultValue = $column['DATA_DEFAULT'] ?? null;
 
         return $c;
@@ -423,43 +478,59 @@ SQL;
 
     /**
      * Finds constraints and fills them into TableSchema object passed.
+     *
      * @param TableSchema $table
+     *
+     * @see https://docs.oracle.com/en/database/oracle/oracle-database/19/refrn/ALL_CONS_COLUMNS.html
+     * @see https://docs.oracle.com/en/database/oracle/oracle-database/19/refrn/ALL_CONSTRAINTS.html
      */
     protected function findConstraints($table)
     {
-        $sql = <<<'SQL'
-SELECT
-    /*+ PUSH_PRED(C) PUSH_PRED(D) PUSH_PRED(E) */
-    D.CONSTRAINT_NAME,
-    D.CONSTRAINT_TYPE,
-    C.COLUMN_NAME,
-    C.POSITION,
-    D.R_CONSTRAINT_NAME,
-    E.TABLE_NAME AS TABLE_REF,
-    F.COLUMN_NAME AS COLUMN_REF,
-    C.TABLE_NAME
-FROM ALL_CONS_COLUMNS C
-    INNER JOIN ALL_CONSTRAINTS D ON D.OWNER = C.OWNER AND D.CONSTRAINT_NAME = C.CONSTRAINT_NAME
-    LEFT JOIN ALL_CONSTRAINTS E ON E.OWNER = D.R_OWNER AND E.CONSTRAINT_NAME = D.R_CONSTRAINT_NAME
-    LEFT JOIN ALL_CONS_COLUMNS F ON F.OWNER = E.OWNER AND F.CONSTRAINT_NAME = E.CONSTRAINT_NAME AND F.POSITION = C.POSITION
-WHERE
-    C.OWNER = :schemaName
-    AND C.TABLE_NAME = :tableName
-ORDER BY D.CONSTRAINT_NAME, C.POSITION
-SQL;
-        $command = $this->db->createCommand($sql, [
-            ':tableName' => $table->name,
-            ':schemaName' => $table->schemaName,
-        ]);
+        $sql = <<<SQL
+            SELECT
+                "d"."CONSTRAINT_NAME",
+                "d"."CONSTRAINT_TYPE",
+                "c"."COLUMN_NAME",
+                "c"."POSITION",
+                "d"."R_CONSTRAINT_NAME",
+                "e"."TABLE_NAME" AS "TABLE_REF",
+                "f"."COLUMN_NAME" AS "COLUMN_REF",
+                "c"."TABLE_NAME"
+            FROM "ALL_CONS_COLUMNS" "c"
+            INNER JOIN "ALL_CONSTRAINTS" "d"
+                ON "d"."OWNER" = "c"."OWNER"
+                AND "d"."CONSTRAINT_NAME" = "c"."CONSTRAINT_NAME"
+            LEFT JOIN "ALL_CONSTRAINTS" "e"
+                ON "e"."OWNER" = "d"."R_OWNER"
+                AND "e"."CONSTRAINT_NAME" = "d"."R_CONSTRAINT_NAME"
+            LEFT JOIN "ALL_CONS_COLUMNS" "f"
+                ON "f"."OWNER" = "e"."OWNER"
+                AND "f"."CONSTRAINT_NAME" = "e"."CONSTRAINT_NAME"
+                AND "f"."POSITION" = "c"."POSITION"
+            WHERE "c"."OWNER" = :schemaName
+                AND "c"."TABLE_NAME" = :tableName
+            ORDER BY "d"."CONSTRAINT_NAME", "c"."POSITION"
+            SQL;
+
+        $command = $this->db->createCommand(
+            $sql,
+            [
+                ':tableName' => $table->name,
+                ':schemaName' => $table->schemaName,
+            ],
+        );
+
         $constraints = [];
+
         foreach ($command->queryAll() as $row) {
-            if ($this->db->slavePdo->getAttribute(\PDO::ATTR_CASE) === \PDO::CASE_LOWER) {
+            if ($this->db->slavePdo->getAttribute(PDO::ATTR_CASE) === PDO::CASE_LOWER) {
                 $row = array_change_key_case($row, CASE_UPPER);
             }
 
             if ($row['CONSTRAINT_TYPE'] === 'P') {
                 $table->columns[$row['COLUMN_NAME']]->isPrimaryKey = true;
                 $table->primaryKey[] = $row['COLUMN_NAME'];
+
                 if (empty($table->sequenceName)) {
                     $table->sequenceName = $this->getTableSequenceName($table->name);
                 }
@@ -472,19 +543,23 @@ SQL;
             }
 
             $name = $row['CONSTRAINT_NAME'];
+
             if (!isset($constraints[$name])) {
                 $constraints[$name] = [
                     'tableName' => $row['TABLE_REF'],
                     'columns' => [],
                 ];
             }
+
             $constraints[$name]['columns'][$row['COLUMN_NAME']] = $row['COLUMN_REF'];
         }
 
         foreach ($constraints as $constraint) {
             $name = current(array_keys($constraint));
-
-            $table->foreignKeys[$name] = array_merge([$constraint['tableName']], $constraint['columns']);
+            $table->foreignKeys[$name] = [
+                $constraint['tableName'],
+                ...$constraint['columns'],
+            ];
         }
     }
 
@@ -502,26 +577,36 @@ SQL;
      * @param TableSchema $table the table metadata
      * @return array all unique indexes for the given table.
      * @since 2.0.4
+     *
+     * @see https://docs.oracle.com/en/database/oracle/oracle-database/19/refrn/ALL_INDEXES.html
+     * @see https://docs.oracle.com/en/database/oracle/oracle-database/19/refrn/ALL_IND_COLUMNS.html
      */
     public function findUniqueIndexes($table)
     {
-        $query = <<<'SQL'
-SELECT
-    DIC.INDEX_NAME,
-    DIC.COLUMN_NAME
-FROM ALL_INDEXES DI
-    INNER JOIN ALL_IND_COLUMNS DIC ON DI.TABLE_NAME = DIC.TABLE_NAME AND DI.INDEX_NAME = DIC.INDEX_NAME
-WHERE
-    DI.UNIQUENESS = 'UNIQUE'
-    AND DIC.TABLE_OWNER = :schemaName
-    AND DIC.TABLE_NAME = :tableName
-ORDER BY DIC.TABLE_NAME, DIC.INDEX_NAME, DIC.COLUMN_POSITION
-SQL;
+        $query = <<<SQL
+            SELECT
+                "dic"."INDEX_NAME",
+                "dic"."COLUMN_NAME"
+            FROM "ALL_INDEXES" "di"
+            INNER JOIN "ALL_IND_COLUMNS" "dic"
+                ON "di"."TABLE_NAME" = "dic"."TABLE_NAME"
+                AND "di"."INDEX_NAME" = "dic"."INDEX_NAME"
+            WHERE "di"."UNIQUENESS" = 'UNIQUE'
+                AND "dic"."TABLE_OWNER" = :schemaName
+                AND "dic"."TABLE_NAME" = :tableName
+            ORDER BY "dic"."TABLE_NAME", "dic"."INDEX_NAME", "dic"."COLUMN_POSITION"
+            SQL;
+
         $result = [];
-        $command = $this->db->createCommand($query, [
-            ':tableName' => $table->name,
-            ':schemaName' => $table->schemaName,
-        ]);
+
+        $command = $this->db->createCommand(
+            $query,
+            [
+                ':tableName' => $table->name,
+                ':schemaName' => $table->schemaName,
+            ],
+        );
+
         foreach ($command->queryAll() as $row) {
             $result[$row['INDEX_NAME']][] = $row['COLUMN_NAME'];
         }
@@ -590,27 +675,39 @@ SQL;
     {
         $params = [];
         $returnParams = [];
+
         $sql = $this->db->getQueryBuilder()->insert($table, $columns, $params);
         $tableSchema = $this->getTableSchema($table);
+
         $returnColumns = $tableSchema->primaryKey;
+
         if (!empty($returnColumns)) {
             $columnSchemas = $tableSchema->columns;
+
             $returning = [];
+
             foreach ((array) $returnColumns as $name) {
                 $phName = QueryBuilder::PARAM_PREFIX . (count($params) + count($returnParams));
+
                 $returnParams[$phName] = [
                     'column' => $name,
                     'value' => '',
                 ];
+
                 if (!isset($columnSchemas[$name]) || $columnSchemas[$name]->phpType !== 'integer') {
                     $returnParams[$phName]['dataType'] = \PDO::PARAM_STR;
                 } else {
                     $returnParams[$phName]['dataType'] = \PDO::PARAM_INT;
                 }
-                $returnParams[$phName]['size'] = isset($columnSchemas[$name]->size) ? $columnSchemas[$name]->size : -1;
+
+                $returnParams[$phName]['size'] = $columnSchemas[$name]->size ?? -1;
                 $returning[] = $this->quoteColumnName($name);
             }
-            $sql .= ' RETURNING ' . implode(', ', $returning) . ' INTO ' . implode(', ', array_keys($returnParams));
+
+            $sql .= ' RETURNING '
+                . implode(', ', $returning)
+                . ' INTO '
+                . implode(', ', array_keys($returnParams));
         }
 
         $command = $this->db->createCommand($sql, $params);
@@ -625,6 +722,7 @@ SQL;
         }
 
         $result = [];
+
         foreach ($returnParams as $value) {
             $result[$value['column']] = $value['value'];
         }
@@ -633,7 +731,32 @@ SQL;
     }
 
     /**
+     * {@inheritdoc}
+     */
+    protected function resolveTableName($name)
+    {
+        $parts = explode('.', str_replace('"', '', $name));
+
+        $tableName = $parts[1] ?? $parts[0];
+
+        $schemaName = isset($parts[1]) ? $parts[0] : $this->defaultSchema;
+
+        $fullName = $schemaName !== $this->defaultSchema
+            ? "{$schemaName}.{$tableName}"
+            : $tableName;
+
+        $resolvedName = new TableSchema();
+
+        $resolvedName->name = $tableName;
+        $resolvedName->schemaName = $schemaName;
+        $resolvedName->fullName = $fullName;
+
+        return $resolvedName;
+    }
+
+    /**
      * Loads multiple types of constraints and returns the specified ones.
+     *
      * @param string $tableName table name.
      * @param string $returnType return type:
      * - primaryKey
@@ -641,80 +764,101 @@ SQL;
      * - uniques
      * - checks
      * @return mixed constraints.
+     *
+     * @see https://docs.oracle.com/en/database/oracle/oracle-database/19/refrn/USER_CONSTRAINTS.html
+     * @see https://docs.oracle.com/en/database/oracle/oracle-database/19/refrn/USER_CONS_COLUMNS.html
      */
     private function loadTableConstraints($tableName, $returnType)
     {
-        static $sql = <<<'SQL'
-SELECT
-    /*+ PUSH_PRED("uc") PUSH_PRED("uccol") PUSH_PRED("fuc") */
-    "uc"."CONSTRAINT_NAME" AS "name",
-    "uccol"."COLUMN_NAME" AS "column_name",
-    "uc"."CONSTRAINT_TYPE" AS "type",
-    "fuc"."OWNER" AS "foreign_table_schema",
-    "fuc"."TABLE_NAME" AS "foreign_table_name",
-    "fuccol"."COLUMN_NAME" AS "foreign_column_name",
-    "uc"."DELETE_RULE" AS "on_delete",
-    "uc"."SEARCH_CONDITION" AS "check_expr"
-FROM "USER_CONSTRAINTS" "uc"
-INNER JOIN "USER_CONS_COLUMNS" "uccol"
-    ON "uccol"."OWNER" = "uc"."OWNER" AND "uccol"."CONSTRAINT_NAME" = "uc"."CONSTRAINT_NAME"
-LEFT JOIN "USER_CONSTRAINTS" "fuc"
-    ON "fuc"."OWNER" = "uc"."R_OWNER" AND "fuc"."CONSTRAINT_NAME" = "uc"."R_CONSTRAINT_NAME"
-LEFT JOIN "USER_CONS_COLUMNS" "fuccol"
-    ON "fuccol"."OWNER" = "fuc"."OWNER" AND "fuccol"."CONSTRAINT_NAME" = "fuc"."CONSTRAINT_NAME" AND "fuccol"."POSITION" = "uccol"."POSITION"
-WHERE "uc"."OWNER" = :schemaName AND "uc"."TABLE_NAME" = :tableName
-ORDER BY "uccol"."POSITION" ASC
-SQL;
+        $sql = <<<SQL
+            SELECT
+                "uc"."CONSTRAINT_NAME" AS "name",
+                "uccol"."COLUMN_NAME" AS "column_name",
+                "uc"."CONSTRAINT_TYPE" AS "type",
+                "fuc"."OWNER" AS "foreign_table_schema",
+                "fuc"."TABLE_NAME" AS "foreign_table_name",
+                "fuccol"."COLUMN_NAME" AS "foreign_column_name",
+                "uc"."DELETE_RULE" AS "on_delete",
+                "uc"."SEARCH_CONDITION" AS "check_expr"
+            FROM "USER_CONSTRAINTS" "uc"
+            INNER JOIN "USER_CONS_COLUMNS" "uccol"
+                ON "uccol"."OWNER" = "uc"."OWNER"
+                AND "uccol"."CONSTRAINT_NAME" = "uc"."CONSTRAINT_NAME"
+            LEFT JOIN "USER_CONSTRAINTS" "fuc"
+                ON "fuc"."OWNER" = "uc"."R_OWNER"
+                AND "fuc"."CONSTRAINT_NAME" = "uc"."R_CONSTRAINT_NAME"
+            LEFT JOIN "USER_CONS_COLUMNS" "fuccol"
+                ON "fuccol"."OWNER" = "fuc"."OWNER"
+                AND "fuccol"."CONSTRAINT_NAME" = "fuc"."CONSTRAINT_NAME"
+                AND "fuccol"."POSITION" = "uccol"."POSITION"
+            WHERE "uc"."OWNER" = :schemaName
+                AND "uc"."TABLE_NAME" = :tableName
+            ORDER BY "uccol"."POSITION" ASC
+            SQL;
 
         $resolvedName = $this->resolveTableName($tableName);
-        $constraints = $this->db->createCommand($sql, [
-            ':schemaName' => $resolvedName->schemaName,
-            ':tableName' => $resolvedName->name,
-        ])->queryAll();
+        $constraints = $this->db->createCommand(
+            $sql,
+            [
+                ':schemaName' => $resolvedName->schemaName,
+                ':tableName' => $resolvedName->name,
+            ],
+        )->queryAll();
         $constraints = $this->normalizePdoRowKeyCase($constraints, true);
         $constraints = ArrayHelper::index($constraints, null, ['type', 'name']);
+
         $result = [
             'primaryKey' => null,
             'foreignKeys' => [],
             'uniques' => [],
             'checks' => [],
         ];
+
         foreach ($constraints as $type => $names) {
             foreach ($names as $name => $constraint) {
                 switch ($type) {
                     case 'P':
-                        $result['primaryKey'] = new Constraint([
-                            'name' => $name,
-                            'columnNames' => ArrayHelper::getColumn($constraint, 'column_name'),
-                        ]);
+                        $result['primaryKey'] = new Constraint(
+                            [
+                                'name' => $name,
+                                'columnNames' => ArrayHelper::getColumn($constraint, 'column_name'),
+                            ],
+                        );
                         break;
                     case 'R':
-                        $result['foreignKeys'][] = new ForeignKeyConstraint([
-                            'name' => $name,
-                            'columnNames' => ArrayHelper::getColumn($constraint, 'column_name'),
-                            'foreignSchemaName' => $constraint[0]['foreign_table_schema'],
-                            'foreignTableName' => $constraint[0]['foreign_table_name'],
-                            'foreignColumnNames' => ArrayHelper::getColumn($constraint, 'foreign_column_name'),
-                            'onDelete' => $constraint[0]['on_delete'],
-                            'onUpdate' => null,
-                        ]);
+                        $result['foreignKeys'][] = new ForeignKeyConstraint(
+                            [
+                                'name' => $name,
+                                'columnNames' => ArrayHelper::getColumn($constraint, 'column_name'),
+                                'foreignSchemaName' => $constraint[0]['foreign_table_schema'],
+                                'foreignTableName' => $constraint[0]['foreign_table_name'],
+                                'foreignColumnNames' => ArrayHelper::getColumn($constraint, 'foreign_column_name'),
+                                'onDelete' => $constraint[0]['on_delete'],
+                                'onUpdate' => null,
+                            ],
+                        );
                         break;
                     case 'U':
-                        $result['uniques'][] = new Constraint([
-                            'name' => $name,
-                            'columnNames' => ArrayHelper::getColumn($constraint, 'column_name'),
-                        ]);
+                        $result['uniques'][] = new Constraint(
+                            [
+                                'name' => $name,
+                                'columnNames' => ArrayHelper::getColumn($constraint, 'column_name'),
+                            ],
+                        );
                         break;
                     case 'C':
-                        $result['checks'][] = new CheckConstraint([
-                            'name' => $name,
-                            'columnNames' => ArrayHelper::getColumn($constraint, 'column_name'),
-                            'expression' => $constraint[0]['check_expr'],
-                        ]);
+                        $result['checks'][] = new CheckConstraint(
+                            [
+                                'name' => $name,
+                                'columnNames' => ArrayHelper::getColumn($constraint, 'column_name'),
+                                'expression' => $constraint[0]['check_expr'],
+                            ],
+                        );
                         break;
                 }
             }
         }
+
         foreach ($result as $type => $data) {
             $this->setTableMetadata($tableName, $type, $data);
         }